More sensible headers
It makes sense to list the audio file in the headers, along with the
original BNC transcripts. It also makes sense to mark the tape sops and
tape starts in a more identifiable way: ((Tape XXXXXX started)) ((Tape
XXXXXX stopped). That should be pretty easy to use in automated
alignments.