Interesting issue with tapes stopping
So here tape 002003 is not manually stopped by Chris, yet the
transcribers mark it as the end of one conversation and the beginning
of the next. This is interesting from a CA perspective as at this point
Chris is *not* choosing to stop the tape, whereas earlier on (in the
transition from tape 020011 to 02002 he was). There is bound to be some
kind of systematic difference between where participants choose to stop
the tape and where transcribers decide that a conversation is over and
needs to be segmented - it certainly doesn’t seem to have anything to
do with the tape medium itself. For the moment, I’ve solved this by
adding another stop-start term ((Tape 002003 not stopped)) (meaning
it’s running on - marked with the last audible few fragments of that
section before talk and transcript begin again in the next tape.

Not an ideal solution, but the best way I can think of sorting it out
for now so I can return to it later.